UniversalDynamics

UniversalDynamics.AbstractMeasure

UniversalDynamics.AbstractNoise

UniversalDynamics.AffineModelDynamics

UniversalDynamics.AffineParameters

UniversalDynamics.BasicFixedIncomeSecurity

UniversalDynamics.ContinuousMoneyMarketAccount

UniversalDynamics.DiagonalNoise

UniversalDynamics.DiscountFactor

UniversalDynamics.DiscreteMoneyMarketAccount

UniversalDynamics.FactorModel

UniversalDynamics.FixedIncomeSecurities

UniversalDynamics.InstantaneousForwardRate

UniversalDynamics.InterestRateModelDynamics

UniversalDynamics.InterestRateModelDynamicsParameters

UniversalDynamics.LiborMarketModelDynamics

UniversalDynamics.LiborMarketModelParameters

UniversalDynamics.MoneyMarketAccount

UniversalDynamics.MultiFactor

UniversalDynamics.MultiFactorAffineModelDynamics

UniversalDynamics.MultiFactorQuadraticModelDynamics

UniversalDynamics.NonDiagonalNoise

UniversalDynamics.OneFactor

UniversalDynamics.OneFactorAffineModelDynamics

UniversalDynamics.OneFactorQuadraticModelDynamics

UniversalDynamics.QuadraticModelDynamics

UniversalDynamics.QuadraticParameters

UniversalDynamics.RiskNeutral

UniversalDynamics.ScalarNoise

UniversalDynamics.ShortRateModelDynamics

UniversalDynamics.ShortRateParameters

UniversalDynamics.SimpleForwardRate

UniversalDynamics.Spot

UniversalDynamics.SpotRate

UniversalDynamics.TForward

UniversalDynamics.TermStructureModelDynamics

UniversalDynamics.Terminal

UniversalDynamics.ZeroCouponBond

UniversalDynamics.q

UniversalDynamics.riccati

UniversalDynamics.riccati!

UniversalDynamics.AbstractMeasure — Type

AbstractMeasure

Supertype for all measures.

UniversalDynamics.AbstractNoise — Type

AbstractNoise{M}

Supertype for noise types with dimension M.

UniversalDynamics.AffineModelDynamics — Type

AffineModelDynamics{FM,IIP,D,DN,T} <: ShortRateModelDynamics{FM,IIP,D,DN,T}

Supertype for Affine Term Structure (ATS) models. These models provide a closed-form formula for the zero coupon bond prices P(t, T), given by:

\[P(t, T) = \exp \left( A(t, T) - B(t, T)^\top \cdot x(t) \right).\]

where $A(t, T)$ and $B(t, T)$ are deterministic functions obtained through a System of Ordinary Differential Equations called Riccati System.

UniversalDynamics.AffineParameters — Type

AffineParameters{FM,D,IIP,DN,K,T,S,A,B,X0,X1,C} <: ShortRateParameters{FM,D,IIP,DN}

Wraps the parameters of an Affine Term Structure (ATS) model in a unique object, for either OneFactor and MultiFactor models.

UniversalDynamics.BasicFixedIncomeSecurity — Type

BasicFixedIncomeSecurity

Supertype for the basic fixed income market securities which are fundamental for Interest Rate modeling.

UniversalDynamics.ContinuousMoneyMarketAccount — Type

ContinuousMoneyMarketAccount <: MoneyMarketAccount

Supertype for continuous money market accounts $eta(t, T)$.

UniversalDynamics.DiagonalNoise — Type

DiagonalNoise{M} <: AbstractNoise{M}

A system of SDEs:

\[d\boldsymbol{x}(t) = d\boldsymbol{μ}(\boldsymbol{x}(t), t) \cdot dt + \dots,\]

has DiagonalNoise if $\boldsymbol{σ}$ is a square diagonal matrix. (igualmente la literatura asume que ademas, la derivada σ[i,i] con respect a x_j es cero. sin embargo esto solo cambia el orden de convergencia del scheme si es que no se cumple).

Aclarar que aqui simplificamos el asunto y usamos un sigma que devuelve un vector ya que es lo mismo que usar la diagonal de una matriz. TODO: Obviamente escribir esto mejor.

UniversalDynamics.DiscountFactor — Type

DiscountFactor <: BasicFixedIncomeSecurity

Supertype for discount factors $D(t, T)$.

UniversalDynamics.DiscreteMoneyMarketAccount — Type

DiscreteMoneyMarketAccount <: MoneyMarketAccount

Supertype for discrete money market accounts $B(t, T)$.

UniversalDynamics.FactorModel — Type

FactorModel

Supertype for factor model types.

UniversalDynamics.FixedIncomeSecurities — Type

FixedIncomeSecurities

Defines an insterest rate object with basic fixed income securities given by an specific interest rate model, such as the ShortRateModel or the LiborMarketModel.

UniversalDynamics.InstantaneousForwardRate — Type

InstantaneousForwardRate <: BasicFixedIncomeSecurity

Supertype for the instantaneous forward rate $f(t, T)$.

UniversalDynamics.InterestRateModelDynamics — Type

InterestRateModelDynamics

Supertype for all Interest Rate Models, such as the ShortRateModel, LiborMarketModel, ForwardMarketModel and Heath-Jarrow-Morton framework.

UniversalDynamics.InterestRateModelDynamicsParameters — Type

InterestRateModelDynamicsParameters

Supertype for all Interest Rate Model Dynamics parameters.

UniversalDynamics.LiborMarketModelDynamics — Type

LiborMarketModelDynamics <: TermStructureModelDynamics

Also known as BGM model, describes... (describir bastante del modelo)

UniversalDynamics.LiborMarketModelParameters — Type

LiborMarketModelParameters

Wraps all the Libor Market Model relevant attributes in a single object.

UniversalDynamics.MoneyMarketAccount — Type

MoneyMarketAccount <: BasicFixedIncomeSecurity

Supertype for money market accounts, also known as bank accounts.

UniversalDynamics.MultiFactor — Type

MultiFactor <: FactorModel

Encompasses all the ShortRateModels where $N$ stochastic factors $x(t)$ determine the future evolution of all interest rates.

UniversalDynamics.MultiFactorAffineModelDynamics — Type

MultiFactorAffineModelDynamics{IIP,D,DN,T,A,P,O} <: AffineModelDynamics{MultiFactor,IIP,D,DN,T}

Defines a MultiFactor ShortRateModelDynamics of Affine type.

UniversalDynamics.MultiFactorQuadraticModelDynamics — Type

MultiFactorQuadraticModelDynamics{IIP,D,DN,T,S,P,O} <: QuadraticModelDynamics{MultiFactor,IIP,D,DN,T}

Defines a MultiFactor ShortRateModel of Quadratic type.

UniversalDynamics.NonDiagonalNoise — Type

NonDiagonalNoise{M} <: AbstractNoise{M}

A System of SDEs has NonDiagonalNoise... tenemos que D != M o que σ no es una matriz diagonal.

\[dX⃗(t) = μ(X⃗(t), t) ⋅ dt + σ(X⃗(t), t) ⋅ dW⃗(t),\]

with $μ$ and $σ$ TODO: Obviamente escribir esto mejor.

UniversalDynamics.OneFactor — Type

OneFactor <: FactorModel

Encompasses all the ShortRateModels where a single stochastic factor $x(t)$ determines the future evolution of all interest rates.

UniversalDynamics.OneFactorAffineModelDynamics — Type

OneFactorAffineModelDynamics{IIP,T,A,P,O} <: AffineModelDynamics{OneFactor,IIP,1,true,T}

Defines a OneFactor ShortRateModelDynamics of Affine type.

UniversalDynamics.OneFactorQuadraticModelDynamics — Type

OneFactorQuadraticModelDynamics{IIP,DN,T,P,O} <: QuadraticModelDynamics{OneFactor,IIP,1,DN,T}

Defines a OneFactor ShortRateModel of Quadratic type.

UniversalDynamics.QuadraticModelDynamics — Type

QuadraticModelDynamics{FM,IIP,D,DN,T} <: ShortRateModelDynamics{FM,IIP,D,DN,T}

Supertype for Quadratic Term Structure (QTS) models. These models provide a closed-form formula for the zero coupon bond prices P(t, T), given by:

\[P(t, T) = \exp \left( -A(t, T) - B(t, T)^\top \cdot x(t) - x(t)^\top \cdot C(t, T) \cdot x(t) \right).\]

where $A(t, T)$, $B(t, T)$ and $C(t, T)$ are deterministic functions obtained through a System of Ordinary Differential Equations called Riccati System.

UniversalDynamics.QuadraticParameters — Type

QuadraticParameters{FM,D,IIP,DN,K,T,S,X0,X1,X2,C} <: ShortRateParameters{FM,D,IIP,DN}

Wraps the parameters of a Quadratic Term Structure (QTS) model in a unique object, for either OneFactor and MultiFactor models.

UniversalDynamics.RiskNeutral — Type

RiskNeutral <: AbstractMeasure

Also known as $Q$, its numeraire is the continuously compounded money market account $\beta(t)$.

UniversalDynamics.ScalarNoise — Type

ScalarNoise <: AbstractNoise{1}

A System of SDEs has ScalarNoise when there is only a unique noise process that affects all the Stochastic Differential Equations. In that sense, the system is given by:

\[dX⃗(t) = μ(X⃗(t), t) ⋅ dt + σ(X⃗(t), t) ⋅ dW(t),\]

with $μ$ and $σ$ functions that returns vectors... and $dW$ a scalar function. TODO: Obviamente escribir esto mejor.

UniversalDynamics.ShortRateModelDynamics — Type

ShortRateModelDynamics{FM,IIP,D,DN,T} <: TermStructureModelDynamics{IIP,D,D,DN,T}

Supertype for ShortRateModels with FactorModel FM, dimension D, in-place or out of place version IIP and with or without diagonal noise DS.

UniversalDynamics.ShortRateParameters — Type

ShortRateParameters{FM,D,IIP,DN}

Supertype for ShortRateModel parameters with FactorModel FM, dimension D, in-place or out of place version IIP and with or without diagonal noise DS.

UniversalDynamics.SimpleForwardRate — Type

SimpleForwardRate <: BasicFixedIncomeSecurity

Supertype for the simply-compounded forward interest rate or simple forward rate $L(t, T, S)$.

UniversalDynamics.Spot — Type

Spot <: AbstractMeasure

Uses the discrete-time equivalent of the continuously compounded money market account as the numeraire. It is often convenient when working with a multitude of forward rates on a tenor structure.

UniversalDynamics.SpotRate — Type

SpotRate <: BasicFixedIncomeSecurity

Supertype for spot rates, also known as short rates.

UniversalDynamics.TForward — Type

TForward{I} <: AbstractMeasure

Also known as $Q^T$, uses a $T$ maturity zero coupon bond $P(t, T)$ as the numeraire. When working with a tenor structure, I identifies $P(t, T_I)$ as the numeraire, with $T_I$ the I-th maturity in the tenor structure.

UniversalDynamics.TermStructureModelDynamics — Type

TermStructureModelDynamics <: InterestRateModelDynamics

Supertype for all Term Structure Interest Rate models.

UniversalDynamics.Terminal — Type

Terminal <: AbstractMeasure

Uses a $T_N$ maturity zero coupon bond $P(t, T_N)$, with $T_N$ the last maturity in the tenor structure, as numeraire.

UniversalDynamics.ZeroCouponBond — Type

ZeroCouponBond <: BasicFixedIncomeSecurity

Supertype for zero coupon bonds, also known as discount bonds, $P(t, T)$.

UniversalDynamics.q — Method

q(Tenors::Vector{<:Real}, t::Real)

Returns the index in the tenor structure Tenors such that Tenors[q(t)-1] ≤ t < Tenors[q(t)], which corresponds to a left continuous interpretation.

UniversalDynamics.riccati! — Method

riccati!(du, u, p::ShortRateParameters, t) -> Nothing

In place computation of a ShortRateModel Riccati's System.

UniversalDynamics.riccati — Method

riccati(u, p::ShortRateParameters, t) -> SVector

Out of place computation of a ShortRateModel Riccati's System.