Introduction

In UniversalDynamics a Dynamics represents continuous time, $D$-dimensional Ito Systems of Stochastic Differential Equations (SDEs) in a time span $\mathbb{I} = \left[ t_0, T \right]$:

\[d\vec{u}(t) = f(t, \vec{u}(t)) \cdot dt + g(t, \vec{u}(t)) \cdot d\vec{W}(t), \quad \vec{u}(t_0) = \vec{u}_0,\\\]

with drift coefficient $f \colon \mathbb{I} \times \mathbb{R}^D \rightarrow \mathbb{R}^D$, diffusion coefficient $g \colon \mathbb{I} \times \mathbb{R}^D \rightarrow \mathbb{R}^{D \times M}$, $M$-dimensional driving Wiener correlated or uncorrelated process $\vec{W}(t)$ and initial condition $\vec{u}_0$.

The main abstract type for all Dynamics is given by:

UniversalDynamics.AbstractDynamics — Type

abstract type AbstractDynamics{InPlace,Dim,NoiseDim,DiagNoise,elType} end

Supertype for all kind of dynamics.

Type parameters:

InPlace: states wether coefficients are in or out of place,

Dim: dynamics dimension,

NoiseDim: noise dimension,

DiagNoise: indicates if the noise is of DiagonalNoise or NonDiagonalNoise type,

It is worth mentioning that there exists simpler cases of Dynamics, which are fairly common, namely:

UniversalDynamics.ScalarNoise — Type

ScalarNoise <: AbstractNoise{1}

A $D$-dimensional system has ScalarNoise when there is only a unique noise process that affects all the Stochastic Differential Equations. In that sense, the system is given by:

\[du⃗(t) = f(t, u⃗(t)) ⋅ dt + g(t, u⃗(t)) ⋅ dW(t), \quad u⃗(t₀) = u⃗₀,\]

in a time span $\mathbb{I} = \left[ t₀, T \right]$ with drift coefficient $f \colon \mathbb{I} × \mathbb{R}ᴰ → \mathbb{R}ᴰ$, diffusion coefficient $g \colon \mathbb{I} × \mathbb{R}ᴰ → \mathbb{R}ᴰ$, $1$-dimensional driving Wiener process $dW(t)$ and initial condition $u⃗₀$.

UniversalDynamics.DiagonalNoise — Type

DiagonalNoise{M} <: AbstractNoise{M}

A $D$-dimensional system has DiagonalNoise when there are $M = D$ noise processes that affect each Stochastic Differential Equation individually. In that sense, the system is given by:

\[du⃗(t) = f(t, u⃗(t)) ⋅ dt + g(t, u⃗(t)) ⋅ d\vec{W}(t), \quad u⃗(t₀) = u⃗₀,\]

in a time span $\mathbb{I} = \left[ t₀, T \right]$ with drift coefficient $f \colon \mathbb{I} × \mathbb{R}ᴰ → \mathbb{R}ᴰ$, diffusion coefficient $g \colon \mathbb{I} × \mathbb{R}^D → \mathrm{diag} \colon \mathbb{R}^{D × D}$, $D$-dimensional driving Wiener correlated or uncorrelated process $d\vec{W}(t)$ and initial condition $u⃗₀$.

In these kind of systems, the diagonal of $g(t, u⃗(t))$ is represented by a $D$-dimensional vector. Then, the product $g(t, u⃗(t)) ⋅ d\vec{W}(t)$ is replaced by the broadcasted product .*.

UniversalDynamics.NonDiagonalNoise — Type

NonDiagonalNoise{M} <: AbstractNoise{M}

A $D$-dimensional system has NonDiagonalNoise when there are M noise processes that affect the Stochastic Differential Equations. In that sense, the system is given by:

\[du⃗(t) = f(t, u⃗(t)) ⋅ dt + g(t, u⃗(t)) ⋅ d\vec{W}(t), \quad u⃗(t₀) = u⃗₀,\]

in a time span $\mathbb{I} = \left[ t₀, T \right]$ with drift coefficient $f \colon \mathbb{I} × \mathbb{R}ᴰ → \mathbb{R}ᴰ$, diffusion coefficient $g \colon \mathbb{I} × \mathbb{R}ᴰ → \mathbb{R}^{D × M}$, $M$-dimensional driving Wiener correlated or uncorrelated process $d\vec{W}(t)$ and initial condition $u⃗₀$.

The library wraps AbstractDynamics attributes inside the following structure:

UniversalDynamics.DynamicsAttributes — Type

DynamicsAttributes

Holds parameters related to any AbstractDynamics.

Fields:

t0: initial time,

x0: initial state,

ρ: correlation matrix,

noise: Wiener process, and

noise_rate_prototype: prototype or representation of the diffusion coefficient.

Representation

Dynamics are represented by two main types, SystemDynamics and ModelDynamics, described below.

SystemDynamics

UniversalDynamics.SystemDynamics — Type

SystemDynamics{IIP,D,M,DN,T} <: AbstractDynamics{IIP,D,M,DN,T}

Represents dynamics with arbitrary coefficients.

Type parameters:

See AbstractDynamics for detailed information.

Fields:

attributes: see DynamicsAttributes for detailed information.

Declaration

SystemDynamics(
    x0::S;
    t0=zero(eltype(S)), ρ::R=I, noise::AbstractNoise=DiagonalNoise{length(x0)}(),
) -> SystemDynamics

returns a SystemDynamics with the given fields, such as state or initial condition x0, intial time t0, correlation matrix ρ and a driving Wiener process noise. Remaining type parameters are obtained through:

IIP: true if isa(x0, Vector) or false if isa(x0, Union{Real,SVector}),

D: equals to length(x0),

M: determined by noise, default value is D,

DN: determined by noise, default value is true.

ModelDynamics

UniversalDynamics.ModelDynamics — Type

abstract type ModelDynamics{IIP,D,M,DN,T} <: AbstractDynamics{IIP,D,M,DN,T} end

Supertype for all dynamics with known coefficients.

ModelDynamics subtypes include many common financial models dynamics. Even though they could always be declared as regular SystemDynamics, it is somewhat useful to have them coded in the library. This enables traceable, reproducible and fast code. Also, for some models, there are many other features implemented in the library, such as Interest Rate Modeling features.

The following is a list with implemented model dynamics:

UniversalDynamics.EquityModelDynamics — Type

abstract type EquityModelDynamics{IIP,D,M,DN,T} <: ModelDynamics{IIP,D,M,DN,T} end

Supertype for all Equity Models dynamics, such as:

BlackScholesMertonModelDynamics,

UniversalDynamics.InterestRateModelDynamics — Type

abstract type InterestRateModelDynamics{IIP,D,M,DN,T} <: ModelDynamics{IIP,D,M,DN,T} end

Supertype for all Interest Rate Models dynamics, such as:

ShortRateModelDynamics,

LiborMarketModelDynamics,

ForwardMarketModelDynamics,

HeathJarrowMortonModelDynamics.

UniversalDynamics.VolatilityModelDynamics — Type

abstract type VolatilityModelDynamics{IIP,D,M,DN,T} <: ModelDynamics{IIP,D,M,DN,T} end

Supertype for all Volatility Models dynamics, such as:

LocalVolatilityModelDynamics,

StochasticVolatilityModelDynamics,

See each corresponding model dynamics section for detailed information.

Examples

Dₓ = 3
Mₓ = 5
x0 = @SVector ones(Dₓ)
x = SystemDynamics(x0; noise=NonDiagonalNoise(Mₓ))

3-dimensional system dynamics with 5-dimensional NonDiagonalNoise
 state:           [1.0, 1.0, 1.0]
 in-place?        false
 Dimension:       3
 Noise dimension: 5
 diagonal noise?  false