Dynamical System

Supose we would like to build a System of SDEs that comes from the union of a given collection of AbstractDynamics, with either arbitrary (see SystemDynamics) or known (see ModelDynamics) coefficients. To be more precise, given a set of Dynamics for $\{ x(t), y(t), z(t) \}$:

\[\begin{aligned} d\vec{x}(t) &= f_x(t, \vec{x}(t)) \cdot dt + g_x(t, \vec{x}(t)) \cdot d\vec{W}_x(t), \quad \vec{x}(t_0) = \vec{x}_0,\\ d\vec{y}(t) &= f_y(t, \vec{y}(t)) \cdot dt + g_y(t, \vec{y}(t)) \cdot d\vec{W}_y(t), \quad \vec{y}(t_0) = \vec{y}_0,\\ d\vec{z}(t) &= f_z(t, \vec{z}(t)) \cdot dt + g_z(t, \vec{z}(t)) \cdot d\vec{W}_z(t), \quad \vec{z}(t_0) = \vec{z}_0,\\ \end{aligned}\]

we wish to obtain the resulting general Dynamics for $u(t)$:

\[d\vec{u}(t) = f(t, \vec{u}(t)) \cdot dt + g(t, \vec{u}(t)) \cdot d\vec{W}(t), \quad \vec{u}(t_0) = \vec{u}_0,\\\]

with:

\[\vec{u}(t) = \begin{bmatrix} \vec{x}(t) \\ \vec{y}(t) \\ \vec{z}(t) \end{bmatrix} \quad f(t, \vec{u}(t)) = \begin{bmatrix} f_x(t, \vec{x}(t)) \\ f_y(t, \vec{y}(t)) \\ f_z(t, \vec{z}(t)) \end{bmatrix} \quad g(t, \vec{u}(t)) = \begin{bmatrix} g_x(t, \vec{x}(t)) & 0 & 0 \\ 0 & g_y(t, \vec{y}(t)) & 0 \\ 0 & 0 & g_z(t, \vec{z}(t)) \end{bmatrix} \quad d\vec{W}(t) = \begin{bmatrix} \vec{W}_x(t) \\ \vec{W}_y(t) \\ \vec{W}_z(t) \end{bmatrix}\]

A DynamicalSystem provides a shorthand for constructing all the previous prototypes which are needed in the Stochastic Differential Equation solvers. However, it does not construct the general functions $f$ and $g$. This task is left to the user. It does provide many useful handlers that come in handy when coding $f$ and $g$.

UniversalDynamics.DynamicalSystem — Type

DynamicalSystem{IIP,D,M,DN,T} <: AbstractDynamics{IIP,D,M,DN,T}

The central structure of UniversalDynamics. Its main purpose is to construct an AbstractDynamics given a collection of AbstractDynamics.

Type parameters:

See AbstractDynamics for detailed information.

Fields:

f: drift coefficient represented as either an in place or out of place function,

g: diffusion coefficient represented as either an in place or out of place function,

attributes:

t0: initial time,

x0: initial state,

ρ: correlation matrix, and

noise: Wiener process.

params: container with parameters values (preferably a NamedTuple),

dynamics: collection of AbstractDynamics (preferably in a OrderedDict),

securities: useful handlers for both dynamics simulation and derivatives pricing.

Declaration:

Given a collection of dynamics, group them together in a container such as an OrderedDict and define a DynamicalSystem using:

DynamicalSystem(dynamics)

The resulting object will have information that can be inspected and is useful for coding the coefficients (drift and difussion) functions.

The complete declaration of a DynamicalSystem requires:

DynamicalSystem(f, g, dynamics, params)

with f and g either the in place or the out of place functions for the coefficients:

Out of place: coefficients must be in the form f(u, p, t) -> SVector and g(u, p, t) -> SVector for DiagonalNoise cases and g(u, p, t) -> SMatrix for the NonDiagonalNoise cases.

In place: coefficients must be in the form f(du, u, p, t) -> nothing and g(du, u, p, t) -> nothing.

Examples

x = SystemDynamics(rand(1); noise=ScalarNoise())
y = SystemDynamics(rand(2); noise=DiagonalNoise(2), ρ=[1 0.3; 0.3 1])
z = SystemDynamics(rand(3); noise=NonDiagonalNoise(2), ρ=[1 0.2; 0.2 1])
dynamics = OrderedDict(:x => x, :y => y, :z => z)
ds = DynamicalSystem(dynamics)

6-dimensional dynamical system
 state:           [0.687848, 0.696561, 0.281215, 0.722819, 0.787821, 0.996618]
 in-place?        true
 Dimension:       6
 Noise dimension: 5
 diagonal noise?  false

Supose we want to price a european option on a stock S with stochastic interest rates (sacar del cap 1 del Andersen). In this context we need to simulate, for example, a short rate described by any ShortRateModelDynamics and a stock price given by a SystemDynamics.

TODO: equations

# define dynamics
x = MultiFactorAffineModelDynamics(x0, ϰ, θ, Σ, α, β, ξ₀, ξ₁)
S = SystemDynamics(S0; noise=NonDiagonalNoise(Mₛ))

# container
dynamics = OrderedDict(:x => x, :S => S)

# define dynamical system formed by the given dynamics
dynamical_system = DynamicalSystem(dynamics)

This will allow the user to check important attributes...

However, in order to solve a DynamicalSystem, the drift f and the diffusion g functions must be provided.