Dynamics

Dynamics represents continuous time, $D$-dimensional Ito Systems of Stochastic Differential Equations (SDEs):

\[d\vec{u}(t) = f(t, \vec{u}(t)) \cdot dt + g(t, \vec{u}(t)) \cdot d\vec{W}(t), \quad \vec{u}(t_0) = \vec{u}_0,\\\]

with drift coefficient $f \colon \left[t_0, T \right] \times \mathbb{R}^D \rightarrow \mathbb{R}^D$, diffusion coefficient $g \colon \left[ t_0, T \right] \times \mathbb{R}^D \rightarrow \mathbb{R}^{D \times M}$, $M$-dimensional driving Wiener correlated or uncorrelated process $d\vec{W}(t)$ and initial condition $\vec{u}_0$.

The main abstract type for these kind of objects is given by:

UniversalDynamics.AbstractDynamics — Type

abstract type AbstractDynamics{InPlace,Dim,NoiseDim,DiagNoise,elType} end

Supertype for all kind of dynamics.

Type parameters:

InPlace: states wether coefficients are in or out of place,

Dim: dynamics dimension,

NoiseDim: noise dimension,

DiagNoise: indicates if the noise is of DiagonalNoise or NonDiagonalNoise type,

It is worth mentioning that there exists simpler cases of Dynamics, which are fairly common, namely:

UniversalDynamics.ScalarNoise — Type

ScalarNoise <: AbstractNoise{1}

A $D$-dimensional system has ScalarNoise when there is only a unique noise process that affects all the Stochastic Differential Equations. In that sense, the system is given by:

\[du⃗(t) = f(t, u⃗(t)) ⋅ dt + g(t, u⃗(t)) ⋅ dW(t), \quad u⃗(t₀) = u⃗₀,\]

with drift coefficient $f \colon \left[ t₀, T \right] × \mathbb{R}ᴰ → \mathbb{R}ᴰ$, diffusion coefficient $g \colon \left[ t₀, T \right] × \mathbb{R}ᴰ → \mathbb{R}ᴰ$, $1$-dimensional driving Wiener process $dW(t)$ and initial condition $u⃗₀$.

UniversalDynamics.DiagonalNoise — Type

DiagonalNoise{M} <: AbstractNoise{M}

A $D$-dimensional system has DiagonalNoise when there are $M = D$ noise processes that affect each Stochastic Differential Equation individually. In that sense, the system is given by:

\[du⃗(t) = f(t, u⃗(t)) ⋅ dt + g(t, u⃗(t)) ⋅ d\vec{W}(t), \quad u⃗(t₀) = u⃗₀,\]

with drift coefficient $f \colon \left[ t₀, T \right] × \mathbb{R}ᴰ → \mathbb{R}ᴰ$, diffusion coefficient $g \colon \left[ t₀, T \right] × \mathbb{R}^D → \mathrm{diag} \colon \mathbb{R}^{D × D}$, $D$-dimensional driving Wiener correlated or uncorrelated process $d\vec{W}(t)$ and initial condition $u⃗₀$.

In these kind of systems, the diagonal of $g(t, u⃗(t))$ is represented by a $D$-dimensional vector. Then, the product $g(t, u⃗(t)) ⋅ d\vec{W}(t)$ is replaced by the broadcasted product .*.

UniversalDynamics.NonDiagonalNoise — Type

NonDiagonalNoise{M} <: AbstractNoise{M}

A $D$-dimensional system has NonDiagonalNoise when there are M noise processes that affect the Stochastic Differential Equations. In that sense, the system is given by:

\[du⃗(t) = f(t, u⃗(t)) ⋅ dt + g(t, u⃗(t)) ⋅ d\vec{W}(t), \quad u⃗(t₀) = u⃗₀,\]

with drift coefficient $f \colon \left[ t₀, T \right] × \mathbb{R}ᴰ → \mathbb{R}ᴰ$, diffusion coefficient $g \colon \left[ t₀, T \right] × \mathbb{R}ᴰ → \mathbb{R}^{D × M}$, $M$-dimensional driving Wiener correlated or uncorrelated process $d\vec{W}(t)$ and initial condition $u⃗₀$.

Dynamics representation

Dynamics are represented by two main types, SystemDynamics and ModelDynamics which are described below.

SystemDynamics

UniversalDynamics.SystemDynamics — Type

SystemDynamics{IIP,D,M,DN,T} <: AbstractDynamics{IIP,D,M,DN,T}

Represents dynamics with arbitrary coefficients.

Type parameters:

See AbstractDynamics for detailed information.

Fields:

t0: initial time,

x0: initial state,

ρ: correlation matrix, and

noise: Wiener process.

Declaration

SystemDynamics(x0::S;
    t0=zero(eltype(S)), ρ::R=I, noise::AbstractNoise=DiagonalNoise{length(x0)}(),
) -> SystemDynamics

returns a SystemDynamics with the given fields, such as state or initial condition x0, intial time t0, correlation matrix ρ and a driving Wiener process noise. Remaining type parameters are obtained through:

IIP: true if isa(x0, Vector) or false if isa(x0, Union{Real,SVector}),

D: equals to length(x0),

M: is determined by noise and its default value is equal to D,

DN: is determined by noise, which is DiagonalNoise{D} by default.

ModelDynamics

UniversalDynamics.ModelDynamics — Type

abstract type ModelDynamics{IIP,D,M,DN,T} <: AbstractDynamics{IIP,D,M,DN,T} end

Supertype for all dynamics with known coefficients.

ModelDynamics subtypes include:

UniversalDynamics.EquityModelDynamics — Type

abstract type EquityModelDynamics{IIP,D,M,DN,T} <: ModelDynamics{IIP,D,M,DN,T} end

Supertype for all equity models dynamics, such as:

BlackScholesMertonModelDynamics,

UniversalDynamics.InterestRateModelDynamics — Type

abstract type InterestRateModelDynamics{IIP,D,M,DN,T} <: ModelDynamics{IIP,D,M,DN,T} end

Supertype for all Interest Rate Models, such as:

ShortRateModelDynamics,

LiborMarketModelDynamics,

ForwardMarketModelDynamics,

HeathJarrowMortonModelDynamics.

UniversalDynamics.VolatilityModelDynamics — Type

abstract type VolatilityModelDynamics{IIP,D,M,DN,T} <: ModelDynamics{IIP,D,M,DN,T} end

Supertype for all volatility models dynamics, such as:

LocalVolatilityModelDynamics,

StochasticVolatilityModelDynamics,

See each corresponding model dynamics for detailed information.