Dynamics

So far we have only mentioned the existence of UniversalDynamics.InterestRateModelDynamics. We now focus on Term Structure Models, given by the following abstract type:

UniversalDynamics.TermStructureModelDynamics — Type

abstract type TermStructureModelDynamics{IIP,D,M,DN,T} <: InterestRateModelDynamics{IIP,D,M,DN,T} end

Supertype for all Term Structure Interest Rate models.

The library implements the following Term Structure Models:

UniversalDynamics.ShortRateModelDynamics — Type

abstract type ShortRateModelDynamics{FM,IIP,D,DN,T} <: TermStructureModelDynamics{IIP,D,D,DN,T} end

Supertype for all Short Rate Models. Notice that the diffusion coefficient for Short Rate Models is a square matrix $∈ \mathbb{R}^{D × D}$, i.e. M = D.

See Short Rate Models section for detailed information.

UniversalDynamics.LiborMarketModelDynamics — Type

LiborMarketModelDynamics{IIP,D,M,DN,T} <: TermStructureModelDynamicsTermStructureModelDynamics{IIP,D,M,DN,T}

Also known as BGM model, describes...

See Libor Market Model section for detailed information.