Dynamical System

In UniversalDynamics a Dynamical System or, more precisely, a Stochastic Dynamical System represents a continuous time, $D$-dimensional Ito System of Stochastic Differential Equations:

\[d\vec{u}(t) = f(t, \vec{u}(t)) \cdot dt + g(t, \vec{u}(t)) \cdot d\vec{W}(t), \quad \vec{u}(t_0) = \vec{u}_0,\\\]

with drift coefficient $f \colon \left[t_0, T \right] \times \mathbb{R}^D \rightarrow \mathbb{R}^D$, diffusion coefficient $g \colon \left[ t_0, T \right] \times \mathbb{R}^D \rightarrow \mathbb{R}^{D \times M}$, $M$-dimensional driving Wiener correlated or uncorrelated process $d\vec{W}(t)$ and initial condition $\vec{u}_0$.

The previous equation represent the most general case of a Dynamical System, which is referenced as the non-diagonal noise case. There are other simpler cases that are really important and are implemented in the library, namely:

UniversalDynamics.ScalarNoise — Type

ScalarNoise <: AbstractNoise{1}

A $D$-dimensional system has ScalarNoise when there is only a unique noise process that affects all the Stochastic Differential Equations. In that sense, the system is given by:

\[du⃗(t) = f(t, u⃗(t)) ⋅ dt + g(t, u⃗(t)) ⋅ dW(t), \quad u⃗(t_0) = u⃗₀,\]

with drift coefficient $f \colon \left[ t₀, T \right] × \mathbb{R}ᴰ → \mathbb{R}ᴰ$, diffusion coefficient $g \colon \left[ t₀, T \right] × \mathbb{R}ᴰ → \mathbb{R}ᴰ$, $1$-dimensional driving Wiener process $dW(t)$ and initial condition $u⃗₀$.

UniversalDynamics.DiagonalNoise — Type

DiagonalNoise{M} <: AbstractNoise{M}

A $D$-dimensional system has DiagonalNoise when there are $M = D$ noise processes that affect each Stochastic Differential Equation individually. In that sense, the system is given by:

\[du⃗(t) = f(t, u⃗(t)) ⋅ dt + g(t, u⃗(t)) ⋅ d\vec{W}(t), \quad u⃗(t_0) = u⃗₀,\]

with drift coefficient $f \colon \left[ t₀, T \right] × \mathbb{R}ᴰ → \mathbb{R}ᴰ$, diffusion coefficient $g \colon \left[ t₀, T \right] × \mathbb{R}^D → \mathrm{diag} \colon \mathbb{R}^{D × D}$, $D$-dimensional driving Wiener correlated or uncorrelated process $d\vec{W}(t)$ and initial condition $u⃗₀$.

In the context of quantitative finance we might want to declare a DynamicalSystem formed by a set of dynamics, with either arbitrary or known coefficients. To take this into account, UniversalDynamics uses the following type architecture:

abstract type AbstractDynamics{InPlace,Dim,NoiseDim,DiagNoise,elType} end

abstract type ModelDynamics{D,M,IIP,DN,T} <: AbstractDynamics{D,M,IIP,DN,T} end

struct SystemDynamics{IIP,D,M,DN,T,A} <: AbstractDynamics{IIP,D,M,DN,T}
    attributes::A
end

UniversalDynamics.SystemDynamics — Type

SystemDynamics{IIP,D,M,DN,T,A} <: AbstractDynamics{IIP,D,M,DN,T}

Represents arbitrary system dynamics with:

In place coefficients if IIP is true, otherwise it is assumed to have out of place coefficients,

Dynamics dimension D,

Noise dimension M,

Diagonal noise if DN is true and non-diagonal noise otherwise,

initial time t0,

initial state x0,

correlation matrix ρ, and

Wiener process noise.

System Dynamics declaration

SystemDynamics(x0::S;
    t0=zero(eltype(S)), ρ::R=I, noise::AbstractNoise=DiagonalNoise{length(x0)}(),
) -> SystemDynamics

returns a SystemDynamics with the given state or initial condition x0, intial time t0, correlation matrix ρ and a driving Wiener process noise. Remaining type parameters are obtained through:

IIP depends on the type of x0 and it is true if isa(x0, Vector) or false if isa(x0, Real) or isa(x0, SVector),

D equals to length(x0),

M is determined by noise and its default value is equal to D,

DN is determined by noise, which is DiagonalNoise by default.

SystemDynamics representing arbitrary dynamics;

ModelDynamics representing known models dynamics, such as:

BlackScholesMerton,

ShortRateModelDynamics,

LiborMarketModelDynamics,

HeathJarrowMortonFrameworkDynamics,

HestonModelDynamics,

...

Dynamical System definition

Supose we want to price a european option on a stock S with stochastic interest rates (sacar del cap 1 del Andersen). In this context we need to simulate, for example, a short rate described by any ShortRateModelDynamics and a stock price given by a SystemDynamics.

TODO: equations

# define dynamics
x = MultiFactorAffineModelDynamics(x0, ϰ, θ, Σ, α, β, ξ₀, ξ₁)
S = SystemDynamics(S0; noise=NonDiagonalNoise(Mₛ))

# container
dynamics = OrderedDict(:x => x, :S => S)

# define dynamical system formed by the given dynamics
dynamical_system = DynamicalSystem(dynamics)

This will allow the user to check important attributes...

However, in order to solve a DynamicalSystem, the drift f and the diffusion g functions must be provided.